% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percent.R
\name{add_percent}
\alias{add_percent}
\title{Title}
\usage{
add_percent(.data, ...)
}
\arguments{
\item{group_vars}{}
}
\value{

}
\description{
Title
}
\examples{
tntpr::wisc \%>\%
count(grade, school) \%>\%
add_percent(grade)
}
