% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_perc_and_ns.R
\name{adorn_perc_and_ns}
\alias{adorn_perc_and_ns}
\title{adorn_perc_and_ns}
\usage{
adorn_perc_and_ns(tabyl, denominator, digits = 0)
}
\arguments{
\item{digits}{}
}
\value{

}
\description{
adorn_perc_and_ns
}
