% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_stacked.R
\name{bar_chart_stack}
\alias{bar_chart_stack}
\title{Make stacked bar chart in TNTP style}
\usage{
bar_chart_stack(df, yvar, xvar, font = "Segoe UI", font_size = 12)
}
\arguments{
\item{df}{data frame}

\item{yvar}{fill variable}

\item{xvar}{group}

\item{font_size}{}
}
\description{
Make stacked bar chart in TNTP style
}
