% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_lm_tidy_w_stars}
\alias{fit_lm_tidy_w_stars}
\title{bar_chart_stacked}
\usage{
fit_lm_tidy_w_stars(dat, formula)
}
\arguments{
\item{dat}{data.frame}

\item{formula}{formula for lm}
}
\value{
data.frame
}
\description{
bar_chart_stacked
}
\examples{
fit_lm_tidy_w_stars(trees, Girth ~ Volume)
}
