% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_tools.R
\name{fix_SM_dual_row_names}
\alias{fix_SM_dual_row_names}
\title{Fix Survey Monkey's dual row names}
\usage{
fix_SM_dual_row_names(dat)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
Fix Survey Monkey's dual row names
}
