% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_formidable_entries.R
\name{pull_formidable_entries}
\alias{pull_formidable_entries}
\title{Pull Formidable Entries}
\usage{
pull_formidable_entries(url, my_username, my_password, page_size = 500)
}
\arguments{
\item{url}{website url}

\item{my_username}{Formidable Username}

\item{my_password}{Formidable Password}
}
\value{
Returns the Formidable entries as a data frame.
}
\description{
Pull Formidable Entries
}
