% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabyl_3way_long.R, R/tabyl_long.R
\name{tabyl_3way_long}
\alias{tabyl_3way_long}
\title{tabyl_3way_long}
\usage{
tabyl_3way_long(tabyl, var3)

tabyl_3way_long(tabyl, var3)
}
\arguments{
\item{var3}{}
}
\value{



}
\description{
tabyl_3way_long

tabyl_3way_long
}
