% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabyl_ms_stacked_bar.R
\name{tabyl_ms_stacked_bar}
\alias{tabyl_ms_stacked_bar}
\title{Data Formatted for MS Stacked Bar Charts}
\usage{
tabyl_ms_stacked_bar(dat, var1, var2)
}
\arguments{
\item{dat}{data}

\item{var1}{row variable}

\item{var2}{column variable}
}
\value{

}
\description{
Data Formatted for MS Stacked Bar Charts
}
