% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tt2_get_query}
\alias{tt2_get_query}
\title{Feed it a filepath where a query lives in a .txt file, it fetches the query}
\usage{
tt2_get_query(path_to_query)
}
\arguments{
\item{path_to_query}{file path}
}
\description{
Feed it a filepath where a query lives in a .txt file, it fetches the query
}
