% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_over_year_movement.R
\name{year_over_year_movement}
\alias{year_over_year_movement}
\title{Teacher Year-Over-Year Movement}
\usage{
year_over_year_movement(year1, year2)
}
\arguments{
\item{year2}{}
}
\value{

}
\description{
Teacher Year-Over-Year Movement
}
