% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_stacked.R
\name{bar_chart_stack}
\alias{bar_chart_stack}
\title{Make stacked bar chart in TNTP style}
\usage{
bar_chart_stack(df, var, group_var, labels = "pct", coord_flip = FALSE,
  title = NULL, xlab = NULL, ylab = NULL, font = "Segoe UI",
  font_size = 8)
}
\arguments{
\item{df}{data frame}

\item{ylab}{}
}
\description{
Make stacked bar chart in TNTP style
}
